Move the filter classes
