+++ app-modules/prospect/src/Filament/Resources/ProspectResource/Pages/ListProspects.php
@@ -7,7 +7,6 @@ use App\Filament\Columns\IdColumn;
 use Filament\Actions\CreateAction;
 use Assist\Prospect\Models\Prospect;
 use App\Filament\Actions\ImportAction;
-use App\Filament\Filters\SelectFilter;
 use Filament\Tables\Actions\EditAction;
 use Filament\Tables\Actions\ViewAction;
 use Filament\Tables\Columns\TextColumn;
@@ -17,6 +16,7 @@ use App\Concerns\FilterTableWithOpenSearch;
 use Filament\Tables\Actions\BulkActionGroup;
 use Assist\Prospect\Imports\ProspectImporter;
 use Filament\Tables\Actions\DeleteBulkAction;
+use App\Filament\Filters\OpenSearch\SelectFilter;
 use OpenSearch\ScoutDriverPlus\Support\Query;
 use OpenSearch\ScoutDriverPlus\Decorators\Hit;
 use Assist\CaseloadManagement\Enums\CaseloadModel;
